pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-protocol"
version = "0.1.0"
description = "The AI that refuses to think for you. Anti-deskilling framework that protects human cognitive skills."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Lorenzo Famiglini", email = "lorenzofamiglini@gmail.com"},
]
keywords = [
    "ai", "cognitive-science", "deskilling", "education",
    "socratic-method", "llm", "hermes-agent", "anti-deskilling",
    "thinking", "metacognition",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Education",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
vertex = [
    "anthropic[vertex]>=0.39.0",
]

[project.urls]
Repository = "https://github.com/lorenzofamiglini/The-Forge-Protocol-Agent"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
include = ["lib*", "plugin*"]
